refactor(utils): introduce per-file rule builder exports
