--- .changeset/chatty-cycles-feel.md
@@ -1,8 +1,0 @@
----
-'eslint-plugin-zod-core': minor
-'eslint-plugin-zod-mini': minor
-'eslint-plugin-zod': minor
-'@eslint-zod/utils': minor
----
-
-refactor: move shared rule implementations into `@eslint-zod/utils`
