+++ .changeset/flat-dogs-eat.md
@@ -1,0 +1,7 @@
+---
+'eslint-plugin-zod': patch
+'eslint-plugin-zod-mini': patch
+'eslint-plugin-zod-core': patch
+---
+
+refactor: update rule builder imports to per-file `@eslint-zod/utils/rule-builders/*` paths
