+++ packages/utils/package.json
@@ -32,6 +32,66 @@
       "import": "./dist/index.mjs",
       "require": "./dist/index.cjs"
     },
+    "./rule-builders/consistent-import": {
+      "@eslint-zod/source": "./src/rule-builders/consistent-import.ts",
+      "import": "./dist/rule-builders/consistent-import.mjs",
+      "require": "./dist/rule-builders/consistent-import.cjs"
+    },
+    "./rule-builders/consistent-import-source": {
+      "@eslint-zod/source": "./src/rule-builders/consistent-import-source.ts",
+      "import": "./dist/rule-builders/consistent-import-source.mjs",
+      "require": "./dist/rule-builders/consistent-import-source.cjs"
+    },
+    "./rule-builders/consistent-object-schema-type": {
+      "@eslint-zod/source": "./src/rule-builders/consistent-object-schema-type.ts",
+      "import": "./dist/rule-builders/consistent-object-schema-type.mjs",
+      "require": "./dist/rule-builders/consistent-object-schema-type.cjs"
+    },
+    "./rule-builders/consistent-schema-output-type-style": {
+      "@eslint-zod/source": "./src/rule-builders/consistent-schema-output-type-style.ts",
+      "import": "./dist/rule-builders/consistent-schema-output-type-style.mjs",
+      "require": "./dist/rule-builders/consistent-schema-output-type-style.cjs"
+    },
+    "./rule-builders/consistent-schema-var-name": {
+      "@eslint-zod/source": "./src/rule-builders/consistent-schema-var-name.ts",
+      "import": "./dist/rule-builders/consistent-schema-var-name.mjs",
+      "require": "./dist/rule-builders/consistent-schema-var-name.cjs"
+    },
+    "./rule-builders/no-any-schema": {
+      "@eslint-zod/source": "./src/rule-builders/no-any-schema.ts",
+      "import": "./dist/rule-builders/no-any-schema.mjs",
+      "require": "./dist/rule-builders/no-any-schema.cjs"
+    },
+    "./rule-builders/no-empty-custom-schema": {
+      "@eslint-zod/source": "./src/rule-builders/no-empty-custom-schema.ts",
+      "import": "./dist/rule-builders/no-empty-custom-schema.mjs",
+      "require": "./dist/rule-builders/no-empty-custom-schema.cjs"
+    },
+    "./rule-builders/no-unknown-schema": {
+      "@eslint-zod/source": "./src/rule-builders/no-unknown-schema.ts",
+      "import": "./dist/rule-builders/no-unknown-schema.mjs",
+      "require": "./dist/rule-builders/no-unknown-schema.cjs"
+    },
+    "./rule-builders/prefer-enum-over-literal-union": {
+      "@eslint-zod/source": "./src/rule-builders/prefer-enum-over-literal-union.ts",
+      "import": "./dist/rule-builders/prefer-enum-over-literal-union.mjs",
+      "require": "./dist/rule-builders/prefer-enum-over-literal-union.cjs"
+    },
+    "./rule-builders/require-brand-type-parameter": {
+      "@eslint-zod/source": "./src/rule-builders/require-brand-type-parameter.ts",
+      "import": "./dist/rule-builders/require-brand-type-parameter.mjs",
+      "require": "./dist/rule-builders/require-brand-type-parameter.cjs"
+    },
+    "./rule-builders/require-error-message": {
+      "@eslint-zod/source": "./src/rule-builders/require-error-message.ts",
+      "import": "./dist/rule-builders/require-error-message.mjs",
+      "require": "./dist/rule-builders/require-error-message.cjs"
+    },
+    "./rule-builders/schema-error-property-style": {
+      "@eslint-zod/source": "./src/rule-builders/schema-error-property-style.ts",
+      "import": "./dist/rule-builders/schema-error-property-style.mjs",
+      "require": "./dist/rule-builders/schema-error-property-style.cjs"
+    },
     "./package.json": "./package.json"
   },
   "publishConfig": {
