+++ packages/utils/src/rule-builders/consistent-object-schema-type.ts
