+++ packages/utils/src/rule-builders/consistent-schema-output-type-style.ts
