+++ packages/utils/src/rule-builders/consistent-schema-var-name.ts
