+++ packages/utils/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown';
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/rule-builders/*.ts'],
   format: ['esm', 'cjs'],
   dts: true,
   clean: true,
