+++ plugins/eslint-plugin-zod-mini/src/rules/consistent-import.ts
@@ -1,9 +1,9 @@
+import { zodMiniImportScope } from '@eslint-zod/utils';
 import {
   IMPORT_SYNTAXES,
   buildConsistentImportCreate,
-  zodMiniImportScope,
-} from '@eslint-zod/utils';
-import type { ImportSyntax } from '@eslint-zod/utils';
+} from '@eslint-zod/utils/rule-builders/consistent-import';
+import type { ImportSyntax } from '@eslint-zod/utils/rule-builders/consistent-import';
 
 import { createZodMiniPluginRule } from '../utils/create-plugin-rule.js';
