+++ plugins/eslint-plugin-zod-mini/src/rules/no-empty-custom-schema.ts
@@ -1,4 +1,5 @@
-import { buildNoEmptyCustomSchemaCreate, zodMiniImportScope } from '@eslint-zod/utils';
+import { zodMiniImportScope } from '@eslint-zod/utils';
+import { buildNoEmptyCustomSchemaCreate } from '@eslint-zod/utils/rule-builders/no-empty-custom-schema';
 
 import { createZodMiniPluginRule } from '../utils/create-plugin-rule.js';
