+++ plugins/eslint-plugin-zod-mini/src/rules/schema-error-property-style.ts
@@ -1,4 +1,5 @@
-import { buildSchemaErrorPropertyStyleCreate, zodMiniImportScope } from '@eslint-zod/utils';
+import { zodMiniImportScope } from '@eslint-zod/utils';
+import { buildSchemaErrorPropertyStyleCreate } from '@eslint-zod/utils/rule-builders/schema-error-property-style';
 
 import { createZodMiniPluginRule } from '../utils/create-plugin-rule.js';
