+++ plugins/eslint-plugin-zod/src/rules/consistent-object-schema-type.ts
@@ -1,4 +1,5 @@
-import { buildConsistentObjectSchemaTypeCreate, zodImportScope } from '@eslint-zod/utils';
+import { zodImportScope } from '@eslint-zod/utils';
+import { buildConsistentObjectSchemaTypeCreate } from '@eslint-zod/utils/rule-builders/consistent-object-schema-type';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
