+++ plugins/eslint-plugin-zod/src/rules/consistent-schema-var-name.ts
@@ -1,4 +1,5 @@
-import { buildConsistentSchemaVarNameCreate, zodImportScope } from '@eslint-zod/utils';
+import { zodImportScope } from '@eslint-zod/utils';
+import { buildConsistentSchemaVarNameCreate } from '@eslint-zod/utils/rule-builders/consistent-schema-var-name';
 
 import { createZodPluginRule } from '../utils/create-plugin-rule.js';
